fully implemented commission fee

+++ server-side/Basketball.Core/Dtos/OrderDto.cs
@@ -8,7 +8,7 @@
         public string? StudentFullName { get; set; }
         public required string TrainingPlanTitle { get; set; }
         public decimal Price { get; set; }
-        public int? CommissionFee { get; set; }
+        public decimal? CommissionFee { get; set; }
         public bool? IsPaid { get; set; }
     }
 }
