+++ server-side/Basketball.Core/Dtos/TrainingPlanDto.cs
@@ -6,10 +6,10 @@
         public required string Title { get; set; }
         public required string Description { get; set; }
         public required string ShortDescription { get; set; }
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         public bool IsActive { get; set; }
         public int Version { get; set; }
-        public required string Coach { get; set; } 
+        public required string Coach { get; set; }
         public List<TrainingPlanSkillDto> Skills { get; set; } = [];
     }
 }
